requirements.txt
Django==1.7
Pillow==2.5.3
South==1.0
argparse==1.2.1
boto==2.35.2
dj-database-url==0.3.0
dj-static==0.0.6
django-autofixture==0.9.1
django-extensions==1.4.0
django-s3direct==0.3.3
djangorestframework==2.4.3
gunicorn==19.1.1
psycopg2==2.5.4
pytz==2014.7
requests==2.4.1
six==1.8.0
static3==0.5.1
validate-email==1.2
wsgiref==0.1.2
